FF-2465 rename flag-evaluation-details.ts to flag-evaluation-details-builder.ts
