+++ src/evaluator.ts
@@ -3,7 +3,7 @@ import {
   AllocationEvaluationCode,
   FlagEvaluationDetails,
   FlagEvaluationDetailsBuilder,
-} from './flag-evaluation-details';
+} from './flag-evaluation-details-builder';
 import { Flag, Shard, Range, Variation, Allocation, Split, VariationType } from './interfaces';
 import { Rule, matchesRule } from './rules';
 import { MD5Sharder, Sharder } from './sharders';
