+++ src/flag-evaluation-details-builder.ts
